CRUD de empleados terminado
Se termina el CRUD de empleados, con las validaciones

+++ api/admin/empleado.php
@@ -181,7 +181,7 @@ if (isset($_GET['action'])) {
                 }
                 break;
             case 'readOne':
-                if (!$empleado->setId_e($_POST['id_empleado'])) {
+                if (!$empleado->setId($_POST['id_empleado'])) {
                     $result['exception'] = 'Usuario incorrecto';
                 } elseif ($result['dataset'] = $empleado->readOne()) {
                     $result['status'] = 1;
@@ -193,34 +193,34 @@ if (isset($_GET['action'])) {
                 break;
             case 'update':
                 $_POST = $empleado->validateForm($_POST);
-                if (!$empleado->setId_e($_POST['id_empleado'])) {
+                if (!$empleado->setId($_POST['id'])) {
                     $result['exception'] = 'Empleado incorrecto';
                 } elseif (!$data = $empleado->readOne()) {
                     $result['exception'] = 'Empleado inexistente';
-                } elseif (!$empleado->setNombre_e($_POST['nombre_empleado'])) {
+                } elseif (!$empleado->setNombre($_POST['nombre'])) {
                     $result['exception'] = 'Nombre incorrectos';
-                } elseif (!$empleado->setApellido_e($_POST['apellido_empleado'])) {
+                } elseif (!$empleado->setApellido($_POST['apellido'])) {
                     $result['exception'] = 'Apellido incorrectos';
-                } elseif (!$empleado->setDUI_e($_POST['DUI_empleado'])) {
+                } elseif (!$empleado->setDUI($_POST['dui'])) {
                     $result['exception'] = 'DUI incorrecto';
-                } elseif (!$empleado->setDireccion_e($_POST['direccion_empleado'])) {
+                } elseif (!$empleado->setCorreo($_POST['email'])) {
                     $result['exception'] = 'Direccion incorrecto';
-                } elseif (!$empleado->setCodigo_e($_POST['codigo_empleado'])) {
+                } elseif (!$empleado->setCelular($_POST['celular'])) {
                     $result['exception'] = 'Codigo incorrecto';
-                } elseif (!$empleado->setTipo_e($_POST['tipo_empleado'])) {
+                } elseif (!$empleado->setTipo($_POST['tipo_empleado'])) {
                     $result['exception'] = 'Tipo empleado incorrecto';
-                } elseif (!is_uploaded_file($_FILES['foto_empleado']['tmp_name'])) {
+                } elseif (!is_uploaded_file($_FILES['foto']['tmp_name'])) {
                     if ($empleado->updateRow($data['foto_empleado'])) {
                         $result['status'] = 1;
                         $result['message'] = 'Empleado modificado correctamente';
                     } else {
                         $result['exception'] = Database::getException();
                     }
-                } elseif (!$empleado->setFoto_e($_FILES['foto_empleado'])) {
+                } elseif (!$empleado->setFoto($_FILES['foto'])) {
                     $result['exception'] = $empleado->getFileError();
                 } elseif ($empleado->updateRow($data['foto_empleado'])) {
                     $result['status'] = 1;
-                    if ($empleado->saveFile($_FILES['foto_empleado'], $empleado->getRuta(), $empleado->getFoto_e())) {
+                    if ($empleado->saveFile($_FILES['foto'], $empleado->getRuta(), $empleado->getFoto())) {
                         $result['message'] = 'Empleado modificado correctamente';
                     } else {
                         $result['message'] = 'Empleado modificado pero no se guardó la imagen';
@@ -232,14 +232,19 @@ if (isset($_GET['action'])) {
             case 'delete':
                 if ($_POST['id_empleado'] == $_SESSION['id_empleado']) {
                     $result['exception'] = 'No se puede eliminar a sí mismo';
-                } elseif (!$empleado->setId_e($_POST['id_empleado'])) {
+                } elseif (!$empleado->setId($_POST['id_empleado'])) {
                     $result['exception'] = 'Empleado incorrecto';
                 } elseif (!$data = $empleado->readOne()) {
                     $result['exception'] = 'Empleado inexistente';
                 } elseif ($empleado->deleteRow()) {
                     $result['status'] = 1;
-                    if ($empleado->deleteFile($empleado->getRuta(), $data['foto_empleado'])) {
-                        $result['message'] = 'Empleado eliminado correctamente';
+                    // Se verifica si la imagen que existe no es la default, para no eliminarla
+                    if(!$data['foto_empleado']=='1.png'){
+                        if ($empleado->deleteFile($empleado->getRuta(), $data['foto_empleado'])) {
+                            $result['message'] = 'Empleado eliminado correctamente';
+                        }else{
+                            $result['message'] = 'Empleado eliminado pero no se borró la imagen';
+                        }
                     } else {
                         $result['message'] = 'Empleado eliminado pero no se borró la imagen';
                     }
